Add elastic-job-lite-executor modules
