+++ elastic-job-lite-executor/elastic-job-lite-simple-executor/src/main/java/org/apache/shardingsphere/elasticjob/lite/simple/job/SimpleJob.java
@@ -7,7 +7,7 @@
  * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,26 +15,20 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.lite.executor.type.impl;
+package org.apache.shardingsphere.elasticjob.lite.simple.job;
 
+import org.apache.shardingsphere.elasticjob.lite.api.job.ElasticJob;
 import org.apache.shardingsphere.elasticjob.lite.api.job.ShardingContext;
-import org.apache.shardingsphere.elasticjob.lite.api.job.config.JobConfiguration;
-import org.apache.shardingsphere.elasticjob.lite.api.job.type.SimpleJob;
-import org.apache.shardingsphere.elasticjob.lite.executor.JobFacade;
-import org.apache.shardingsphere.elasticjob.lite.executor.type.ClassedJobItemExecutor;
 
 /**
- * Simple job executor.
+ * Simple job.
  */
-public final class SimpleJobExecutor implements ClassedJobItemExecutor<SimpleJob> {
+public interface SimpleJob extends ElasticJob {
     
-    @Override
-    public void process(final SimpleJob elasticJob, final JobConfiguration jobConfig, final JobFacade jobFacade, final ShardingContext shardingContext) {
-        elasticJob.execute(shardingContext);
-    }
-    
-    @Override
-    public Class<SimpleJob> getElasticJobClass() {
-        return SimpleJob.class;
-    }
+    /**
+     * Execute job.
+     *
+     * @param shardingContext sharding context
+     */
+    void execute(ShardingContext shardingContext);
 }
